[입력]수정: Input 클래스 분할
